Refactor logic of pageheader
