+++ src/layouts/PageHeader.tsx
@@ -1,13 +1,10 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useDispatch } from 'react-redux';
-import { AppDispatch, RootState } from '../store/store';
-import { logout } from '../store/auth/authSlice';
+import { RootState } from '../store/store';
 import ProfileDropDown from '../components/ProfileDropDown';
 import { useSelector } from 'react-redux';
 
 const PageHeader = () => {
-  const dispatch = useDispatch<AppDispatch>();
   const [isDropDownOpen, setIsDropDownOpen] = React.useState(false);
   const navigate = useNavigate();
   const isLoggedIn = useSelector((state: RootState) => state.auth.isLoggedIn);
@@ -16,16 +13,15 @@ const PageHeader = () => {
     navigate('/auth/login');
   }
 
-  const handleLogout = () => {
-    const confirmLogout = window.confirm('Are you sure you want to logout?');
-    if (confirmLogout) {
-      dispatch(logout());
-    }
-  }
-
   const toggleDropdown = () => {
     setIsDropDownOpen(!isDropDownOpen);
   };
+
+  const handleProfileClick = () => {
+    isLoggedIn ? toggleDropdown() : handleNavigateToLogin();
+  }
+
+  const profileTitle = isLoggedIn ? 'Profile' : 'Login';
 
   return (
     <div className='sticky top-0 z-20 flex gap-10 lg:gap-20 justify-between 
@@ -40,40 +36,22 @@ const PageHeader = () => {
           Connecting Conversations, Spreading Stories
         </p>
       </div>
-
-      {!isLoggedIn && (
-        <div className='flex items-center'>
-          <div className='relative container'>
-            <img
-              src="/blankprofile.png"
-              alt="Profile"
-              className="w-10 h-10 object-cover rounded-full cursor-pointer"
-              onClick={handleNavigateToLogin}
-              title='You are logged out. Please log in.'
-            />
-          </div>
-        </div>
-      )}
-
-      {
-        isLoggedIn && (
-          <div className='flex items-center'>
-            <div className='relative container'>
-              <img
-                src="/blankprofile.png"
-                alt="Profile"
-                className="w-10 h-10 object-cover rounded-full cursor-pointer"
-                onClick={toggleDropdown}
-              />
-              {isDropDownOpen && (
-                <div>
-                  < ProfileDropDown handleLogout={handleLogout} />
-                </div>
-              )}
+      <div className='flex items-center'>
+        <div className='relative container'>
+          <img
+            src="/blankprofile.png"
+            alt="Profile"
+            className="w-10 h-10 object-cover rounded-full cursor-pointer"
+            onClick={handleProfileClick}
+            title={profileTitle}
+          />
+          {isDropDownOpen && (
+            <div>
+              < ProfileDropDown setIsDropDownOpen={setIsDropDownOpen} />
             </div>
-          </div>
-        )
-      }
+          )}
+        </div>
+      </div>
     </div >
   )
 }
